toynn/act/requirements.txt
# ACT-UNet Requirements
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0
tqdm>=4.65.0
scipy>=1.7.0
scikit-image>=0.19.0

# Optional for advanced features
tensorboard>=2.11.0
wandb>=0.13.0
